src/recommender/requirements.txt
torch>=1.9.0
numpy>=1.19.2
faiss-cpu>=1.7.2  # 如果有GPU可以使用faiss-gpu
oracledb>=1.3.1
typing-extensions>=4.0.0 